Copy qmldir files into build tree with associated plugins.
The qdeclarativemoduleplugin test verifies that the plugins it
constructs are correctly imported.  Ths requires that the qmldir
files that describe the plugins are copied into the build tree
along with the plugins they refer to.

QML (and .js, where relevant) files must also be copied into
the same directory if they form part of the imported module.

Change-Id: I827a0b678a17cac42aadc4f2176fdd8bd693aab1
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>