Windows: Fix qdeclarativemoduleplugin test
The failure was caused by drive letter being wrong case.
The whole QVERIFY here was pointless, as all it was doing was verifying
what QTest::qFindTestData() returned earlier when resolving the imports
directory (path variable), which is in no way relevant to this test -
especially as the drive letter case is anyway fixed on the next line.

Removed the unnecessary QVERIFY and added equivalent if statement to
fix the drive letter case only if it is wrong.

Task-number: QTBUG-24471
Change-Id: I15c9fdd5c451723e7e877709280d7483fc769c32
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>
Reviewed-by: Kai Koehne <kai.koehne@nokia.com>