Fix QDeclarativeItem::hasActiveFocus().
This function returns true if the item (or, in case it's a focus
scope, one of its children,) has focus (i.e., will
receive key events now), or will receive focus once the scene is
activated. It also returns true if the item has not yet been added
to a scene, but has subFocus, implicating that someone has called
setFocus() on it prior to it being added to the scene; the latter
case seen most commonly in unit tests.

Task-number: QTBUG-24681

Change-Id: Ia1265158108d5d856b57e34a8efee2545cd0deeb
Reviewed-by: Andreas Aardal Hanssen <andreas@hanssen.name>
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>
Reviewed-by: Girish Ramakrishnan <girish.1.ramakrishnan@nokia.com>