Do not let QML warnings reduce the frame rate in fullscreen mode.
QPlainTextEdit::appendPlainText() can take surprisingly long, even
when the widget is hidden. In a demo on embedded hardware we've seen
the frame time go up from ~55 milliseconds to ~85 over the first few
minutes of the demo running. There was one warning emitted per frame.
With this patch there is no significant performance degradation.

Change-Id: Ic80409b29a1544f03431e1573811a00925660338
Reviewed-by: Matthew Vogt <matthew.vogt@nokia.com>