Eliminated usage of "declarative" compat module.
The check in qmltooling attempts to avoid installing the qtquick1
version of the debugger plugin if qtquick2 is used.  The removal of the
declarative module meant that this plugin was always enabled,
potentially overwriting the qtquick2 debugger plugin.  This check was
updated to work again, although this approach is still somewhat
problematic.  It would be better if the different versions of the
plugins had different names.

The check in tools/designer/src/plugins doesn't seem to make sense, so
it was removed.

Change-Id: I55d1bf59264d3c5d62c2d4545e9ced952e460efb
Reviewed-by: Kent Hansen <kent.hansen@nokia.com>