Fix memory leak in qdeclarativevisualdatamodel.
Cherry-pick of b26a588c1702a84ba853df3c757d879cc9f2fc46 from 4.8

Parent order was backwards, the inner context was parented to the
outer context, and then the inner context was reparented to the
delegate. Parent the outer context to inner context so both are
deleted when the delegate is deleted.

Task-number: QTBUG-25784

Change-Id: I4a500b41b02d3cb5dfead4b87ac86e9583024108
Reviewed-by: Martin Jones <martin.jones@nokia.com>