Fix install path of folderlistmodel particles and shader import plugins
19bba2faca1e1a1e6ddbb7a899a6894fc64c1723 added QtQuick1 to both the
base path of QtQuick 1 imports and the target paths of individual
plugins leading to a duplication of QtQuick1 in the final path.

Remove the suffix from the root destination directory as the plugins
still need to explicity include it in the install path and it's better
if there's a single relative target path used in the .pro files.

Task-number: QTBUG-26707 QTBUG-26701
Change-Id: I79ee28fc5b6ef549cf3c5f71585756149f0f8ff3
Reviewed-by: Matthew Vogt <matthew.vogt@nokia.com>