Fix install paths of demos and examples.
Install all demos and examples to the qtquick1 subdirectory of
the examples install location, demos under a demos folder,
declarative examples under a declarative folder and the common
viewer code under a helper folder.

This mirrors the directory structure of the src directory and
ensures the examples sources can be built from the install
location.

Task-number: QTBUG-26663
Change-Id: I406f4fa58e5f004eb1aef5902fa2d6267517943d
Reviewed-by: Alan Alpert <alan.alpert@nokia.com>