Accept missing semicolon after do-while statement.
Although not valid according to ECMA 5.1, both JSC and V8 accept a
missing semicolon after a do-while loop. (Both JSC and V8 do not do this
through automatic semicolon insertion, but handle it as a special case.)

This is a "port" of a similair change done in the qtdeclarative module.

Change-Id: I417cf18abb75e7040c48894b8f8f68afc41547fe
Reviewed-by: Lars Knoll <lars.knoll@digia.com>