Observe GL context loss in ShaderEffectItem
Recreate the shader program as required when the GL context
changes. Also delete the shader program when the ShaderEffectItem
is deactivated, to save graphics memory.

Previously, this was done only for Symbian platforms, and in a
suboptimal fashion. This patch implements the behaviour for all
platforms in a more optimal manner.

Task-number: QTBUG-28136
Change-Id: I8281e77f7285ea0046bdf092ab827cae3b00a6d2
Reviewed-by: Robin Burchell <robin+qt@viroteck.net>
Reviewed-by: Alan Alpert <aalpert@rim.com>