Ensure QDeclarativeTextLayout uses the correct pen
When a styled text specifies a font region with some color set,
QDeclarativeTextLayout used to revert to using a default-constructed
(black-colored) pen for text occurring after the </font> tag, rather
than the "default" (property-defined) color.

This commit ensures that the correct pen is used, by setting the pen
of the "inert" painter to be an invalid color rather than a default
constructed pen.  It then detects when the current pen transitions
back to this invalid pen, and switches back to using the default
color.

Task-number: QTBUG-28135
Change-Id: I6bcc43fbcf7fb2c680959d27b1422364ebb473ae
Reviewed-by: Alan Alpert <aalpert@rim.com>
Reviewed-by: Robin Burchell <robin+qt@viroteck.net>