Add a method that allows registration of files to types
There is currently no way in C++ to duplicate the functionality of a qmldir
file in mapping QML files to versioned types in a module. This functionality
would be useful both in cases where a separate qmldir file would be overkill,
and for cases where the type mapping should be generated dynamically.

Change-Id: I28d7898122c5556fcd7cf3476795bcf4bb288eab
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>