Process imports in the order they appear in the code.
By placing imports into a hash keyed by the qualifier, imports were not
processed in the order they appeared in the source. This is perhaps not a major
problem, but it is nice to have predictability.

This becomes more important in Qt 5, when QHash is randomly seeded at
application startup, as alterations to order could have strange side effects on
code.

Reported-by: Maciej Blomberg
Change-Id: Ibd34fab1eed4bc727b498ba4684c62d801ae713b
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>