Improve emission of FocusScope focusChanged signals.
Emit activeFocusChanged up the focus tree when an item receives a focus
event, rather than when the focusItem changes which happens before
internal state is sufficiently updated for hasFocus and hasActiveFocus
to return the correct values from within a changed signal handler.

There are some limitions to this. First the signals are not emitted
reliably when the scene is not active which makes sense for activeFocus
but not for focus.  The second is in some instances the focus and
activeFocus can update unnecessarily while the focus chain sorts itself
out.

QDeclarativeItem tests by Andreas Aardal Hanssen <andreas@hanssen.name>

Task-number: QTBUG-28288
Task-number: QTBUG-25644

Change-Id: Ib3d17c42754c15a08b34c3388f50b45cc1d2a831
Reviewed-by: Andreas Aardal Hanssen <andreas@hanssen.name>
Reviewed-by: Alan Alpert <aalpert@rim.com>