Fix regression from Qt 4 for deep group properties
Setting a property like someObject.font.family would cause an out-of-bounds
assertion in QList, called from QQmlCompiler::genValueTypeProperty. This
appears to be a regression from commit
5e970b84663f5398eb51d4575856d1a3c44df953, which replaced one -1 to
QMetaType::QVariant too many times. It appears the use of -1 is rather
deliberate here and not to indicate the use of a QVariant property.

The attached test verifies this as well as the successful setting of the
property at the end.

Task-number: QTBUG-31576
Change-Id: I237ea08847e1db31481a311ea8ec23a5ccc702d8
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Andrew den Exter <andrew.den.exter@qinetic.com.au>