Correct QDeclarativeTextEdit::positionToRectangle().
d->yoff is the content offset when the content is not top aligned.
This means that positionToRectangle() returned incorrect values when
the content was bottom or vertical center aligned. In other words,
it didn't apply the offset but gave coordinates from the beginning of
the content instead of from the beginning of the item.

Change-Id: I044b79434b9a67b0539050e39953d5889a3ed1db
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>