QDeclarativeError::toString: improve handling of empty urls.
"file::2:23: ..." is strange to read. Show "<Unknown File>:2:23: ..." instead,
by treating empty urls (including "file:") as unknown, and by still showing
line and column numbers in such a case.

This change makes it possible for QUrl::fromLocalFile("") to return an
empty url rather than "file:", which this module was relying upon in the tests.

Change-Id: Id6d8aaa73673283cb65cbd1316dca77f859a3f8f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>