Fix Flickable mouse grab issue.
This issue cannot be reproduced on the desktop with a mouse (at least I
could not).  It can however reliably be reproduced on iMX6 embedded hardware
with a touch-screen.  It seems like the TS driver is reporting mouse events
with a much more coarse granularity than the USB mouse driver, which in turn
triggers this bug.
(Qt4's declarative, as well as Qt5's quick1 and 2 all have the same issue)

Change-Id: I01c9fc433a78d5f79518786783e24f3f93eaa50a
Task-number: QTBUG-27185
Reviewed-by: Martin Jones <martin.jones@jollamobile.com>