Fix autotest to not rely on broken operator<<(QDebug, QUrl) behavior
QmlInfo does call nospace() in the constructor. However, the
streaming operator for QDebug di permanently reset this to space(),
which the autotest relies on. This got fixed in qtbase change
868201155fd, letting this autotest fail.

Move QUrl to the end of the chain so that the broken behavior of
QUrl isn't checked anymore.

This is a cherry-pick of change 2aec854665f62e2177 in qtdeclarative.

Change-Id: I40f775749392e1c7516b9316706976862025f806
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>