Skip failing autotest tst_qdeclarativetextedit::delegateLoading
The test failures point to a race condition in QtNetwork when a socket
is prematurely closed by the server. Anyhow, the race condition has
been there before, just that a recent change in Qt5Core made it more
likely to be hit.

Disabling the test for now: This is better handled in qtbase, and
shouldn't block the release.

Task-number: QTBUG-48299
Change-Id: If6877d47940d422e94b6d20d24a8fbdb89692f71
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>